refactor loader
