+++ tests/unit/config/test_loader.py
@@ -10,6 +10,8 @@ from toolbelt.config import ToolbeltConfig
 from toolbelt.config import loader as loader_mod
 from toolbelt.config.loader import (
     load_config,
+)
+from toolbelt.config.file_loaders import (
     load_config_from_file,
     load_python_config,
     load_yaml_config,
